+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 * MLOPs Nodes can now specify input paths as dictionaries if default input configs are known.
 * `coco_align` no longer handles annotations
+* MLOps can now handle input nodes with multiple connections (variable length inputs)
 
 
 ## Version 0.6.8 - Target 2023-05-22